refactor(generator): 移除代码生成模块相关代码
- 删除 generator-web、generator-server、generator-domain 中所有代码生成相关的控制器、服务与引擎实现
- 删除 generator-web 中 application.yml 配置文件
- 清理代码生成相关的 VO、枚举及仓库定义
- 彻底移除代码生成模块的功能及依赖，简化项目结构

feat(literule): 增强审批记录数据结构及方法

- 在 ApprovalRecord 中添加委托人 delegateFrom 字段
- 新增记录创建时间和更新时间字段
- 新增 currentLevelApprovedApprovers 列表缓存和对应的访问方法
- 实现追加审批日志 appendLog 方法
- 提供依据审批级别查询已审批人员列表方法
- 实现审批状态判断方法 isStatus

fix(auth): 增加用户名称 HTTP 头兼容性常量

- 添加 X_USER_NAME 作为 X_USERNAME 的别名，兼容 nextwiki/fileComment 等模块

feat(aicontroller): 使用 SummaryResult 替换内部 DTO 类型

- AiController 中将 generateSummary 方法返回类型由 NextwikiDto.SummaryResult 修改为 SummaryResult
- 调用 aiSummaryService.generateSummary 时直接返回 SummaryResult 类型，简化代码

refactor(liteexpr): 完善字节码编译器设计与字节码解释器兼容性

- BytecodeCompiler 实现完整的 AST 到字节码编译流程，支持多种表达式类型及逻辑短路
- 引入常量池管理，提高编译效率及字节码模块化
- 优化跳转指令管理，支持跳转偏移回填
- BytecodeInterpreter 调整常量池元素访问类型，统一为 String，移除 VarName 包装类
- 增强字节码解释的稳定性和兼容性

style(liteexpr): 优化 BytecodeCompiler 的包和导入语句

- 删除无用注释，调整导包顺序和代码格式
- 扩大字节码列表初始容量，提高性能

+++ ydsz-literule/ydsz-literule-server/src/main/java/com/njydsz/literule/server/config/RuleAdminService.java
@@ -278,7 +278,7 @@ public class RuleAdminService {
     def.setReviewedBy(vo.getReviewedBy());
     def.setReviewedAt(vo.getReviewedAt());
     def.setReviewComment(vo.getReviewComment());
-    def.setCanaryRatio(vo.getCanaryRatio() != null ? vo.getCanaryRatio() : BigDecimal.ZERO);
+    def.setCanaryRatio(vo.getCanaryRatio() != null ? BigDecimal.valueOf(vo.getCanaryRatio()) : BigDecimal.ZERO);
     def.setCanaryConditionExpression(vo.getCanaryConditionExpression());
     def.setCanarySeverityExpression(vo.getCanarySeverityExpression());
     return def;
