+++ ydsz-literule/ydsz-literule-server/src/main/java/com/njydsz/literule/server/config/RuleSearchService.java
@@ -131,7 +131,7 @@ public class RuleSearchService {
     def.setReviewedBy(vo.getReviewedBy());
     def.setReviewedAt(vo.getReviewedAt());
     def.setReviewComment(vo.getReviewComment());
-    def.setCanaryRatio(vo.getCanaryRatio() != null ? vo.getCanaryRatio() : BigDecimal.ZERO);
+    def.setCanaryRatio(vo.getCanaryRatio() != null ? BigDecimal.valueOf(vo.getCanaryRatio()) : BigDecimal.ZERO);
     def.setCanaryConditionExpression(vo.getCanaryConditionExpression());
     def.setCanarySeverityExpression(vo.getCanarySeverityExpression());
     return def;
